feat: Add `-svg` flag
