+++ CHANGELOG.adoc
@@ -14,6 +14,12 @@ All notable changes to this project will be documented in this file.
 The format is based on https://keepachangelog.com/[Keep a Changelog], and this
 project adheres to https://semver.org/[Semantic Versioning].
 
+== {compare-url}/v0.1.1\...HEAD[Unreleased]
+
+=== Added
+
+* Add `-svg` flag ({pull-request-url}/6[#6])
+
 == {compare-url}/v0.1.0\...v0.1.1[0.1.1] - 2025-05-14
 
 === Added
