+++ assets/output.svg.license
@@ -1,0 +1,3 @@
+SPDX-FileCopyrightText: 2025 Shun Sakai
+
+SPDX-License-Identifier: CC0-1.0
