+++ cli.go
@@ -14,6 +14,7 @@ type options struct {
 	width     uint
 	height    uint
 	blockSize uint
+	svg       bool
 	version   bool
 }
 
@@ -23,6 +24,7 @@ func init() {
 	flag.UintVar(&opt.width, "width", 256, "Set the width of pixel art")
 	flag.UintVar(&opt.height, "height", 256, "Set the height of pixel art")
 	flag.UintVar(&opt.blockSize, "block-size", 1, "Set the block size of pixel art")
+	flag.BoolVar(&opt.svg, "svg", false, "Generate SVG file instead of PNG file")
 	flag.BoolVar(&opt.version, "version", false, "Print version number")
 
 	flag.Usage = func() {
