+++ main.go
@@ -7,15 +7,8 @@ package main
 import (
 	"flag"
 	"fmt"
-	"image"
-	"image/color"
-	"image/png"
 	"log"
-	"math"
-	"math/rand/v2"
 	"os"
-	"path/filepath"
-	"sync"
 )
 
 func main() {
@@ -31,45 +24,16 @@ func main() {
 		os.Exit(1)
 	}
 
-	outputPath := filepath.Clean(flag.Arg(0))
-
+	outputPath := flag.Arg(0)
 	width, height, blockSize := int(opt.width), int(opt.height), int(opt.blockSize)
-	img := image.NewNRGBA(image.Rect(0, 0, width, height))
 
-	var wg sync.WaitGroup
-	for y := 0; y < height; y += blockSize {
-		for x := 0; x < width; x += blockSize {
-			wg.Add(1)
-			go func(x, y int) {
-				defer wg.Done()
-				r := uint8(rand.N(math.MaxUint8))
-				g := uint8(rand.N(math.MaxUint8))
-				b := uint8(rand.N(math.MaxUint8))
-				a := uint8(math.MaxUint8)
-
-				for by := 0; (by < blockSize) && (y+by < height); by++ {
-					for bx := 0; (bx < blockSize) && (x+bx < width); bx++ {
-						img.Set(x+bx, y+by, color.NRGBA{r, g, b, a})
-					}
-				}
-			}(x, y)
-		}
-	}
-
-	wg.Wait()
-
-	output, err := os.Create(outputPath)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	defer func() {
-		if err := output.Close(); err != nil {
+	if opt.svg {
+		if err := generateSVG(outputPath, width, height, blockSize); err != nil {
 			log.Fatal(err)
 		}
-	}()
-
-	if err := png.Encode(output, img); err != nil {
-		log.Print(err)
+	} else {
+		if err := generatePNG(outputPath, width, height, blockSize); err != nil {
+			log.Fatal(err)
+		}
 	}
 }
