+++ testdata/help.ct
@@ -8,6 +8,8 @@ Usage: pagen [OPTIONS] <FILE>
     	Set the block size of pixel art (default 1)
   -height uint
     	Set the height of pixel art (default 256)
+  -svg
+    	Generate SVG file instead of PNG file
   -version
     	Print version number
   -width uint
@@ -19,6 +21,8 @@ Usage: pagen [OPTIONS] <FILE>
     	Set the block size of pixel art (default 1)
   -height uint
     	Set the height of pixel art (default 256)
+  -svg
+    	Generate SVG file instead of PNG file
   -version
     	Print version number
   -width uint
